@@ -1,60 +1,38 @@
-apply plugin: 'com.android.library'
-
-//ext {
-//  bintrayRepo = 'maven'
-//  bintrayName = 'expandablecheckrecyclerview'
-//
-//  publishedGroupId = 'com.thoughtbot'
-//  libraryName = 'expandablecheckrecyclerview'
-//  artifact = 'expandablecheckrecyclerview'
-//
-//  libraryDescription =
-//      'Custom Android RecyclerViewAdapters that collapse and expand and support checking and unchecking children'
-//
-//  siteUrl = 'https://github.com/thoughtbot/expandable-recycler-view'
-//  gitUrl = 'https://github.com/thoughtbot/expandable-recycler-view.git'
-//
-//  libraryVersion = '1.5'
-//
-//  developerId = 'mandybess'
-//  developerName = 'Amanda Hill'
-//  developerEmail = 'amandabesshill@gmail.com'
-//
-//  licenseName = 'The Apache Software License, Version 2.0'
-//  licenseUrl = 'http://www.apache.org/licenses/LICENSE-2.0.txt'
-//  allLicenses = ["Apache-2.0"]
-//}
+plugins {
+    id 'com.android.library'
+}
 
 android {
-  compileSdkVersion 23
-  buildToolsVersion "23.0.3"
+    namespace 'com.thoughtbot.expandablecheckrecyclerview'
 
-  defaultConfig {
-    minSdkVersion 16
-    targetSdkVersion 23
-    versionCode 5
-    versionName "1.5"
-  }
-  buildTypes {
-    release {
-      minifyEnabled false
+    compileSdk 34
+
+    defaultConfig {
+        minSdkVersion 23
+        targetSdkVersion 34
+
+        versionCode 1
+        versionName "1.0.0"
     }
-  }
+
+    buildTypes {
+        release {
+            minifyEnabled false
+
+            def version = defaultConfig.versionName
+            archivesBaseName = "expandable-check-recycle-view.$version"
+        }
+    }
 }
 
 dependencies {
-  compile fileTree(dir: 'libs', include: ['*.jar'])
-  compile 'com.thoughtbot:expandablerecyclerview:1.3'
+    api "com.google.android.material:material:1.12.0"
+    api "androidx.appcompat:appcompat:1.7.0"
 
-  //android
-  compile 'com.android.support:appcompat-v7:23.4.0'
-  compile 'com.android.support:recyclerview-v7:23.4.0'
+    // unit tests
+    testImplementation 'junit:junit:4.12'
+    testImplementation 'org.robolectric:robolectric:3.2.1'
+    testImplementation 'org.mockito:mockito-core:1.10.5'
 
-  //unit tests
-  testCompile 'junit:junit:4.12'
-  testCompile 'org.robolectric:robolectric:3.2.1'
-  testCompile 'org.mockito:mockito-core:1.10.5'
+    api project(':expandablerecyclerview')
 }
-
-//apply from: 'https://raw.githubusercontent.com/nuuneoi/JCenter/master/installv1.gradle'
-//apply from: 'https://raw.githubusercontent.com/nuuneoi/JCenter/master/bintrayv1.gradle'
