@@ -1,18 +1,23 @@
-apply plugin: 'com.android.application'
+plugins {
+  id "com.android.application"
+}
 
 android {
-  compileSdkVersion 23
-  buildToolsVersion "23.0.3"
+  namespace 'com.thoughtbot.expandablerecyclerview.sample'
+  compileSdk 34
 
   defaultConfig {
     applicationId "com.thoughtbot.expandablerecyclerview.sample"
-    minSdkVersion 16
-    targetSdkVersion 23
+
+    minSdkVersion 23
+    targetSdkVersion 34
+
     versionCode 1
-    versionName "1.0"
+    versionName "1.0.0"
 
     testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
   }
+
   buildTypes {
     release {
       minifyEnabled false
@@ -22,23 +27,22 @@ android {
 }
 
 dependencies {
-  compile project(':expandablecheckrecyclerview')
+  implementation project(':expandablecheckrecyclerview')
 
-  //android
-  compile 'com.android.support:appcompat-v7:23.4.0'
-  compile 'com.android.support:recyclerview-v7:23.4.0'
+  implementation "com.google.android.material:material:1.12.0"
+  implementation "androidx.appcompat:appcompat:1.7.0"
 
   //unit tests
-  testCompile 'junit:junit:4.12'
-  testCompile 'org.robolectric:robolectric:3.2.1'
-  testCompile 'org.mockito:mockito-core:1.10.5'
+  testImplementation 'junit:junit:4.12'
+  testImplementation 'org.robolectric:robolectric:3.2.1'
+  testImplementation 'org.mockito:mockito-core:1.10.5'
 
   //automation tests
-  androidTestCompile 'com.android.support.test.espresso:espresso-core:2.2.2'
-  androidTestCompile 'com.android.support.test:runner:0.5'
-  androidTestCompile 'com.android.support:support-annotations:23.4.0'
-  androidTestCompile 'com.android.support.test:rules:0.4.1'
-  androidTestCompile('com.android.support.test.espresso:espresso-contrib:2.2') {
+  androidTestImplementation 'com.android.support.test.espresso:espresso-core:2.2.2'
+  androidTestImplementation 'com.android.support.test:runner:0.5'
+  androidTestImplementation 'com.android.support:support-annotations:23.4.0'
+  androidTestImplementation 'com.android.support.test:rules:0.4.1'
+  androidTestImplementation('com.android.support.test.espresso:espresso-contrib:2.2') {
     // Necessary to avoid version conflicts
     exclude group: 'com.android.support', module: 'appcompat'
     exclude group: 'com.android.support', module: 'support-v4'
