@@ -1,7 +1,9 @@
 package com.thoughtbot.expandablerecyclerview.viewholders;
 
-import android.support.v7.widget.RecyclerView;
 import android.view.View;
+
+import androidx.recyclerview.widget.RecyclerView;
+
 import com.thoughtbot.expandablerecyclerview.models.ExpandableGroup;
 
 /**
