@@ -1,35 +1,35 @@
 package com.thoughtbot.expandablerecyclerview.sample;
 
 import com.thoughtbot.expandablerecyclerview.models.ExpandableGroup;
+
 import java.util.List;
 
 public class Genre extends ExpandableGroup<Artist> {
 
-  private int iconResId;
+    private final int iconResId;
 
-  public Genre(String title, List<Artist> items, int iconResId) {
-    super(title, items);
-    this.iconResId = iconResId;
-  }
+    public Genre(String title, List<Artist> items, int iconResId) {
+        super(title, items);
+        this.iconResId = iconResId;
+    }
 
-  public int getIconResId() {
-    return iconResId;
-  }
+    public int getIconResId() {
+        return iconResId;
+    }
 
-  @Override
-  public boolean equals(Object o) {
-    if (this == o) return true;
-    if (!(o instanceof Genre)) return false;
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (!(o instanceof Genre)) return false;
 
-    Genre genre = (Genre) o;
+        Genre genre = (Genre) o;
 
-    return getIconResId() == genre.getIconResId();
+        return getIconResId() == genre.getIconResId();
 
-  }
+    }
 
-  @Override
-  public int hashCode() {
-    return getIconResId();
-  }
+    @Override
+    public int hashCode() {
+        return getIconResId();
+    }
 }
-
