@@ -3,9 +3,9 @@ package com.thoughtbot.expandablerecyclerview.sample;
 import android.app.Activity;
 import android.content.Intent;
 import android.os.Bundle;
-import android.view.View;
 import android.view.View.OnClickListener;
 import android.widget.Button;
+
 import com.thoughtbot.expandablerecyclerview.sample.expand.ExpandActivity;
 import com.thoughtbot.expandablerecyclerview.sample.multicheck.MultiCheckActivity;
 import com.thoughtbot.expandablerecyclerview.sample.multitype.MultiTypeActivity;
@@ -14,34 +14,31 @@ import com.thoughtbot.expandablerecyclerview.sample.singlecheck.SingleCheckActiv
 
 public class MainActivity extends Activity {
 
-  @Override
-  protected void onCreate(Bundle savedInstanceState) {
-    super.onCreate(savedInstanceState);
-    setContentView(R.layout.activity_main);
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
 
-    Button expand = (Button) findViewById(R.id.expand_button);
-    expand.setOnClickListener(navigateTo(ExpandActivity.class));
+        Button expand = findViewById(R.id.expand_button);
+        expand.setOnClickListener(navigateTo(ExpandActivity.class));
 
-    Button multiSelect = (Button) findViewById(R.id.multi_check_button);
-    multiSelect.setOnClickListener(navigateTo(MultiCheckActivity.class));
+        Button multiSelect = findViewById(R.id.multi_check_button);
+        multiSelect.setOnClickListener(navigateTo(MultiCheckActivity.class));
 
-    Button singleSelect = (Button) findViewById(R.id.single_check_button);
-    singleSelect.setOnClickListener(navigateTo(SingleCheckActivity.class));
+        Button singleSelect = findViewById(R.id.single_check_button);
+        singleSelect.setOnClickListener(navigateTo(SingleCheckActivity.class));
 
-    Button mixedSelect = (Button) findViewById(R.id.mixedtype_button);
-    mixedSelect.setOnClickListener(navigateTo(MultiTypeActivity.class));
+        Button mixedSelect = findViewById(R.id.mixedtype_button);
+        mixedSelect.setOnClickListener(navigateTo(MultiTypeActivity.class));
 
-    Button mixedTypeAndCheck = (Button) findViewById(R.id.mixedtype_check_button);
-    mixedTypeAndCheck.setOnClickListener(navigateTo(MultiTypeCheckGenreActivity.class));
-  }
+        Button mixedTypeAndCheck = findViewById(R.id.mixedtype_check_button);
+        mixedTypeAndCheck.setOnClickListener(navigateTo(MultiTypeCheckGenreActivity.class));
+    }
 
-  public OnClickListener navigateTo(final Class<?> clazz) {
-    return new OnClickListener() {
-      @Override
-      public void onClick(View v) {
-        Intent intent = new Intent(MainActivity.this, clazz);
-        startActivity(intent);
-      }
-    };
-  }
+    public OnClickListener navigateTo(final Class<?> clazz) {
+        return v -> {
+            Intent intent = new Intent(MainActivity.this, clazz);
+            startActivity(intent);
+        };
+    }
 }
