@@ -2,19 +2,20 @@ package com.thoughtbot.expandablerecyclerview.sample.expand;
 
 import android.view.View;
 import android.widget.TextView;
+
 import com.thoughtbot.expandablerecyclerview.sample.R;
 import com.thoughtbot.expandablerecyclerview.viewholders.ChildViewHolder;
 
 public class ArtistViewHolder extends ChildViewHolder {
 
-  private TextView childTextView;
+    private final TextView childTextView;
 
-  public ArtistViewHolder(View itemView) {
-    super(itemView);
-    childTextView = (TextView) itemView.findViewById(R.id.list_item_artist_name);
-  }
+    public ArtistViewHolder(View itemView) {
+        super(itemView);
+        childTextView = itemView.findViewById(R.id.list_item_artist_name);
+    }
 
-  public void setArtistName(String name) {
-    childTextView.setText(name);
-  }
+    public void setArtistName(String name) {
+        childTextView.setText(name);
+    }
 }
