@@ -2,19 +2,19 @@ package com.thoughtbot.expandablerecyclerview.sample.multicheck;
 
 import com.thoughtbot.expandablecheckrecyclerview.models.MultiCheckExpandableGroup;
 import com.thoughtbot.expandablerecyclerview.sample.Artist;
+
 import java.util.List;
 
 public class MultiCheckGenre extends MultiCheckExpandableGroup {
 
-  private int iconResId;
+    private final int iconResId;
 
-  public MultiCheckGenre(String title, List<Artist> items, int iconResId) {
-    super(title, items);
-    this.iconResId = iconResId;
-  }
+    public MultiCheckGenre(String title, List<Artist> items, int iconResId) {
+        super(title, items);
+        this.iconResId = iconResId;
+    }
 
-  public int getIconResId() {
-    return iconResId;
-  }
+    public int getIconResId() {
+        return iconResId;
+    }
 }
-
