@@ -2,20 +2,20 @@ package com.thoughtbot.expandablerecyclerview.sample.multitype;
 
 import android.view.View;
 import android.widget.TextView;
+
 import com.thoughtbot.expandablerecyclerview.sample.R;
 import com.thoughtbot.expandablerecyclerview.viewholders.ChildViewHolder;
 
 public class FavoriteArtistViewHolder extends ChildViewHolder {
 
-  private TextView favoriteArtistName;
+    private final TextView favoriteArtistName;
 
-  public FavoriteArtistViewHolder(View itemView) {
-    super(itemView);
-    favoriteArtistName = (TextView) itemView.findViewById(R.id.list_item_favorite_artist_name);
-  }
+    public FavoriteArtistViewHolder(View itemView) {
+        super(itemView);
+        favoriteArtistName = itemView.findViewById(R.id.list_item_favorite_artist_name);
+    }
 
-  public void setArtistName(String name) {
-    favoriteArtistName.setText(name);
-  }
-
+    public void setArtistName(String name) {
+        favoriteArtistName.setText(name);
+    }
 }
