@@ -1,42 +1,45 @@
 package com.thoughtbot.expandablerecyclerview.sample.multitype;
 
-import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
-import com.thoughtbot.expandablerecyclerview.sample.R;
-
 import static com.thoughtbot.expandablerecyclerview.sample.GenreDataFactory.makeGenres;
+
+import android.os.Bundle;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
+
+import com.thoughtbot.expandablerecyclerview.sample.R;
 
 public class MultiTypeActivity extends AppCompatActivity {
 
-  private MultiTypeGenreAdapter adapter;
+    private MultiTypeGenreAdapter adapter;
 
-  @Override
-  protected void onCreate(@Nullable Bundle savedInstanceState) {
-    super.onCreate(savedInstanceState);
-    setContentView(R.layout.activity_multi_type);
-    getSupportActionBar().setDisplayHomeAsUpEnabled(true);
-    getSupportActionBar().setTitle(getClass().getSimpleName());
+    @Override
+    protected void onCreate(@Nullable Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_multi_type);
+        getSupportActionBar().setDisplayHomeAsUpEnabled(true);
+        getSupportActionBar().setTitle(getClass().getSimpleName());
 
-    RecyclerView recyclerView = (RecyclerView) findViewById(R.id.recycler_view);
-    LinearLayoutManager layoutManager = new LinearLayoutManager(this);
+        RecyclerView recyclerView = (RecyclerView) findViewById(R.id.recycler_view);
+        LinearLayoutManager layoutManager = new LinearLayoutManager(this);
 
-    adapter = new MultiTypeGenreAdapter(makeGenres());
-    recyclerView.setLayoutManager(layoutManager);
-    recyclerView.setAdapter(adapter);
-  }
+        adapter = new MultiTypeGenreAdapter(makeGenres());
+        recyclerView.setLayoutManager(layoutManager);
+        recyclerView.setAdapter(adapter);
+    }
 
-  @Override
-  protected void onSaveInstanceState(Bundle outState) {
-    super.onSaveInstanceState(outState);
-    adapter.onSaveInstanceState(outState);
-  }
+    @Override
+    protected void onSaveInstanceState(@NonNull Bundle outState) {
+        super.onSaveInstanceState(outState);
+        adapter.onSaveInstanceState(outState);
+    }
 
-  @Override
-  protected void onRestoreInstanceState(Bundle savedInstanceState) {
-    super.onRestoreInstanceState(savedInstanceState);
-    adapter.onRestoreInstanceState(savedInstanceState);
-  }
+    @Override
+    protected void onRestoreInstanceState(Bundle savedInstanceState) {
+        super.onRestoreInstanceState(savedInstanceState);
+        adapter.onRestoreInstanceState(savedInstanceState);
+    }
 }
