@@ -1,53 +1,49 @@
 package com.thoughtbot.expandablerecyclerview.sample.singlecheck;
 
-import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
-import android.view.View;
-import android.view.View.OnClickListener;
-import android.widget.Button;
-import com.thoughtbot.expandablerecyclerview.sample.R;
-
 import static com.thoughtbot.expandablerecyclerview.sample.GenreDataFactory.makeSingleCheckGenres;
+
+import android.os.Bundle;
+import android.widget.Button;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
+
+import com.thoughtbot.expandablerecyclerview.sample.R;
 
 public class SingleCheckActivity extends AppCompatActivity {
 
-  private SingleCheckGenreAdapter adapter;
+    private SingleCheckGenreAdapter adapter;
 
-  @Override
-  protected void onCreate(@Nullable Bundle savedInstanceState) {
-    super.onCreate(savedInstanceState);
-    setContentView(R.layout.activity_single_check);
-    getSupportActionBar().setDisplayHomeAsUpEnabled(true);
-    getSupportActionBar().setTitle(getClass().getSimpleName());
+    @Override
+    protected void onCreate(@Nullable Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_single_check);
+        getSupportActionBar().setDisplayHomeAsUpEnabled(true);
+        getSupportActionBar().setTitle(getClass().getSimpleName());
 
-    RecyclerView recyclerView = (RecyclerView) findViewById(R.id.recycler_view);
-    LinearLayoutManager layoutManager = new LinearLayoutManager(this);
+        RecyclerView recyclerView = (RecyclerView) findViewById(R.id.recycler_view);
+        LinearLayoutManager layoutManager = new LinearLayoutManager(this);
 
-    adapter = new SingleCheckGenreAdapter(makeSingleCheckGenres());
-    recyclerView.setLayoutManager(layoutManager);
-    recyclerView.setAdapter(adapter);
+        adapter = new SingleCheckGenreAdapter(makeSingleCheckGenres());
+        recyclerView.setLayoutManager(layoutManager);
+        recyclerView.setAdapter(adapter);
 
-    Button clear = (Button) findViewById(R.id.clear_button);
-    clear.setOnClickListener(new OnClickListener() {
-      @Override
-      public void onClick(View v) {
-        adapter.clearChoices();
-      }
-    });
-  }
+        Button clear = findViewById(R.id.clear_button);
+        clear.setOnClickListener(v -> adapter.clearChoices());
+    }
 
-  @Override
-  protected void onSaveInstanceState(Bundle outState) {
-    super.onSaveInstanceState(outState);
-    adapter.onSaveInstanceState(outState);
-  }
+    @Override
+    protected void onSaveInstanceState(@NonNull Bundle outState) {
+        super.onSaveInstanceState(outState);
+        adapter.onSaveInstanceState(outState);
+    }
 
-  @Override
-  protected void onRestoreInstanceState(Bundle savedInstanceState) {
-    super.onRestoreInstanceState(savedInstanceState);
-    adapter.onRestoreInstanceState(savedInstanceState);
-  }
+    @Override
+    protected void onRestoreInstanceState(Bundle savedInstanceState) {
+        super.onRestoreInstanceState(savedInstanceState);
+        adapter.onRestoreInstanceState(savedInstanceState);
+    }
 }
