@@ -3,45 +3,47 @@ package com.thoughtbot.expandablerecyclerview.sample.singlecheck;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
 import com.thoughtbot.expandablecheckrecyclerview.CheckableChildRecyclerViewAdapter;
 import com.thoughtbot.expandablecheckrecyclerview.models.CheckedExpandableGroup;
 import com.thoughtbot.expandablerecyclerview.models.ExpandableGroup;
 import com.thoughtbot.expandablerecyclerview.sample.Artist;
 import com.thoughtbot.expandablerecyclerview.sample.R;
 import com.thoughtbot.expandablerecyclerview.sample.expand.GenreViewHolder;
+
 import java.util.List;
 
 public class SingleCheckGenreAdapter extends
-    CheckableChildRecyclerViewAdapter<GenreViewHolder, SingleCheckArtistViewHolder> {
+        CheckableChildRecyclerViewAdapter<GenreViewHolder, SingleCheckArtistViewHolder> {
 
-  public SingleCheckGenreAdapter(List<SingleCheckGenre> groups) {
-    super(groups);
-  }
+    public SingleCheckGenreAdapter(List<SingleCheckGenre> groups) {
+        super(groups);
+    }
 
-  @Override
-  public SingleCheckArtistViewHolder onCreateCheckChildViewHolder(ViewGroup parent, int viewType) {
-    View view = LayoutInflater.from(parent.getContext())
-        .inflate(R.layout.list_item_singlecheck_arist, parent, false);
-    return new SingleCheckArtistViewHolder(view);
-  }
+    @Override
+    public SingleCheckArtistViewHolder onCreateCheckChildViewHolder(ViewGroup parent, int viewType) {
+        View view = LayoutInflater.from(parent.getContext())
+                .inflate(R.layout.list_item_singlecheck_arist, parent, false);
+        return new SingleCheckArtistViewHolder(view);
+    }
 
-  @Override
-  public void onBindCheckChildViewHolder(SingleCheckArtistViewHolder holder, int position,
-      CheckedExpandableGroup group, int childIndex) {
-    final Artist artist = (Artist) group.getItems().get(childIndex);
-    holder.setArtistName(artist.getName());
-  }
+    @Override
+    public void onBindCheckChildViewHolder(SingleCheckArtistViewHolder holder, int position,
+                                           CheckedExpandableGroup group, int childIndex) {
+        final Artist artist = (Artist) group.getItems().get(childIndex);
+        holder.setArtistName(artist.getName());
+    }
 
-  @Override
-  public GenreViewHolder onCreateGroupViewHolder(ViewGroup parent, int viewType) {
-    View view = LayoutInflater.from(parent.getContext())
-        .inflate(R.layout.list_item_genre, parent, false);
-    return new GenreViewHolder(view);
-  }
+    @Override
+    public GenreViewHolder onCreateGroupViewHolder(ViewGroup parent, int viewType) {
+        View view = LayoutInflater.from(parent.getContext())
+                .inflate(R.layout.list_item_genre, parent, false);
+        return new GenreViewHolder(view);
+    }
 
-  @Override
-  public void onBindGroupViewHolder(GenreViewHolder holder, int flatPosition,
-      ExpandableGroup group) {
-    holder.setGenreTitle(group);
-  }
+    @Override
+    public void onBindGroupViewHolder(GenreViewHolder holder, int flatPosition,
+                                      ExpandableGroup group) {
+        holder.setGenreTitle(group);
+    }
 }
